main: comment out the sqlite store of imported side chain blocks
Since the SQLite tables use height as primary key (doh), the side chain
blocks should not be added as they will overwrite the main chain blocks.
In the future, the SQLite tables should be modified to make hash the
primary key (or a dedicated row id) and both is_mainchain and is_valid
added to the blocks table. For now, do not insert into SQLite during
side chain import.