dcrpg: synchronous project update should come outside of sync loop
With the project fund update being a blocking call, it should be called
after the ChainDB sync loop to minimize the possibility of a wiredDB's
sync ending while the balance update runs AND a new block arrives.
The effect will be ChainDB requesting stakeDB connect notification,
but waiting forever since wiredDB (the stakeDB manager) has already
quit.
TODO: Fix this in a way to guarantee no race.