dcrpg: major optimization with pool status update
The pool status update for unspent expired and missed tickets updated
rows by matching hash.  However, the row IDs (primary keys) of these
ticket hashes are available in the TicketTxnIDGetter, the
unspentTicketCache of ChainDB.  This change uses the cache (without
expiring the entries since they may be needed if the tickets are
eventually spent via revocation) so the update does not require scanning
on the (unindexed) hash column.

The above optimization is noticable as there is no loner a significant
slowdown of the block processing further in the process.

The other optimization relates to the new check in UpdateLastBlock
that ensures the previous block is on the same chain.  This edit
replaces BlockStatus with the new BlockFlags query that does not involve
a join but only provides isValid and isMainchain.  However, for main
chain blocks, even the call to BlockStatus is skipped as there should
never be a main chain block with a side chain parent.

make cache miss log a Tracef