cache address and value for funding tx outs
The insertion of spending addresses rows requires the address and value
from the previous outpoint.  This introduces a cache to prevent the need
to query the vouts table for this information.
The cache is part of ChainDB, and is passed to InsertSpendingAddressRow
(and insertSpendingAddressRow) as a pointer. When the pointer is nil,
a DB query gets the address and value, otherwise the passed data is
for the addresses table spending row insertion.

Introduce utxoSore type to manage concurrent access to cache.