main: run final sync just before starting monitors
The final node sync check needs to be run just before starting the
monitor goroutines, but after dcrdata registers for dcrd notifications.
The monitor goroutines start receiving immediately after sync.

This also runs explore.RetrieveUpdates (a slow pre-populate of charts
data) after the monitors are running.  Having it between the final
ensureSync and starting the monitors caused missed blocks, especially
on slower machines.

NOTE: This is still not perfect since there needs to be a sync option
that only goes to a specified height, so that a notification for a
duplicate block cannot happen.  The time window for this to happen is

explorer: send final sigSyncStatus when disabling sync status

RetrieveUpdates only pre-populates charts data now.

RetrieveUpdates --> PrepareCharts