sync: continually insert spending txns, only batch updating funding txns (#754)
This modifies sync to do insertion of address rows for vins (spending 
transactions) along with all the other insertions rather than in a batch 
after syncing all blocks. Only the update to the funding transaction 
rows of the addresses table with their matching tx (the spending tx) is 
done in a batch process at the end, after indexing.

Note that the Block Sync stage is slower since more insertions are being 
performed that were part of a the Batch addresses (table) Update stage. 
However, the Batch Update stage is much quicker since the insertions at 
this stage were being done with indexes present, and queries were 
required to look up the data for the address and value columns.

The performance difference may be much more pronounced on slower 
machines, especially those without NVMe SSDs backing the DB.

Prior to these changes, only the funding address table rows 
(corresponding to vouts) were inserted during the block sync loop, while 
the spending address table rows (corresponding to vins) were inserted 
after the blocks were synchronized. This batch insert of spending rows 
took place at the same time as the updates to the funding tx rows with 
info for any matched spending transaction.

These edits remove insertSpendingTxByPrptStmt, which handled both the 
insertion and updates in the above described batch, and implement a 
separate insert function, InsertSpendingAddressRow, with a flag 
indicating if the funding row updates should be done in addition to the 
spending row insertions (updateFundingRow). When performing the initial 
sync, where a batch update would take place at the end of block 
processing, updateFundingRow=false to only do the insertions. When 
indexes exist and no batch updates are to take place (i.e. normal catch 
up mode when restarting dcrdata), updateFundingRow=true, to do the 
insertions and updates. The updates are now handled in a separate 
function, setSpendingForFundingOP.

This approach was not previously feasible because inserting the spending 
rows required querying for the address and value, by hash. Since there 
is no index initially, a slow scan on hash was required. To avoid this 
query, a utxoStore type is implemented that provides a cache of these 
addresses and values, keyed by outpoint (tx hash : tx out index). This 
data is stored in the cache when new funding transactions are processed 
(those that create the outputs that would otherwise need to be queried 
for). The data is then retrieved from the cache when processing the new 
vins that form the basis for the inserted spending addresses table rows. 
A query is only performed on a cache miss, which may happen in certain 
rare situations.

StoreBlock now reports address rows inserted, in addition to vin and 
vout counts.

SetSpendingForVinDbIDs is also slightly accelerated by the new 
SelectVinVoutPairByID query that gets only the data needed by 
setSpendingForFundingOP (the funding tx outpoint).

This also fixes an issue with the monitors possibly starting without having
the chain fully sync'd.