explorer: wrap NewBlockData and ExtraInfo with mutex
Previously, the NewBlockDataMtx mutex guarded both NewBlockData
and ExtraInfo.  It was also recently overloaded to guard
displaySyncStatusPage.  It was not self evident what data the mutex was
meant to guard.

This creates a pageData struct that embeds a sync.RWMutex, and has
BlockInfo (the previous NewBlockData type) and HomeInfo (the previous
Extra info type).

The displaySyncStatusPage field is changed from a bool to and
atomic.Value.  Access to this variable should be done via
DisplaySyncStatusPage and SetDisplaySyncStatusPage to ensure the
proper type is saved/loaded with the atomic.Value.

MeanVotingBlocks is now computed in the explorerUI constructor and
stored as a field of explorerUI to prevent excessive mutex locking and
recomputing the value.

In Store, simulateASR is no longer computed in the critical section.

PoolInfo.Target is reset on each Store in case PoolInfo is not available
or becomes available.

In storeMempoolInfo, locking of MempoolData was happening much sooner
than necessary.  The Lock is moved to just before needed, and
FormattedTotalSize is generated outside of the critical section.