create a context for shutdown signalling
Create the context and start a listener in main, and pass the context to
_main, where it is passed to other goroutines and functions. This
replaces the "quit" channel. Instead of close(quit), main calls
requestShutdown(), which signals shutdownListener to cancel the context
created by withShutdownCancel.

All of the long-running sync functions, monitor goroutines, etc. are now
plumbed with the context where previously they used the quit chan.

With the context now a field of ChainDB, the next step should be to pass
the context to the postgres queries.  After that, child contexts with
timeouts should be created for each query. This will allow both
cancellation and timeouts.

The http handlers should also have access to the context.