Header footer fixup (#785)
This fixes the logo not showing, and restores header/footer formatting.

It also removes the header and logo related css from nexthome.css.
So the heights of header and footer are back to their previous dimensions.
The footer markup has been reworded a bit so that it reflows better.
The logo now collapses to just the symbol for narrow screens.
Also, there's now a white version of logo for night mode and sim/test net modes.