Mempool transactions on /tx update (mostly) correctly now (#777)
Maturity is set to "Mature" when a ticket matures.
Progress bars get live updates, and are restyled.

The only difference between a confirmed transaction's /tx page and a
mempool /tx page which is confirmed after page/turbolinks load is that
the "Pool Status" is not displayed after a mempool tx is confirmed.