Disable buttons that plot a chart with one record. (#798)
For addresses with 1 or no data points, this disables the Charts button.

This also disables inapplicable Zoom and Group By buttons.

Under the hood, there is some refactoring:
This adds the dbtypes.AddressMetrics type.
This adds the SelectAddressTimeGroupingCount query, which gets the number of groups
for a given address and time grouping interval.
(*ChainDB).GetOldestTxBlockTime is replaced with GetAddressMetrics. Similarly,
the explorer.explorerDataSource is modified for this change.
AddressPageData replaces OldestTxTime with Metrics.
In controllers/address.js, setApplicableBtns now takes an txCountByType parameter
so it can decide which buttons to disable (when txCountByType<2).
Instead of id, Stimulus data-oldestblockTime is used for the oldest block time.