Elided text, vout/vin index, and table-layout: auto on tx page tables. (#778)
This center-elides text in txid and address columns on the tx page.

This also adds the vout/vin index to the In and Out tables.

Pure-CSS middle-eliding.

This makes changes to the table layout on the transactions page. The abridged/elided
text is created using the template.FuncMap, so it can be used on any other page too.

This also removes the signature script from coinbase and stakebase inputs as they
are of little interest and never change.