Additional websocket updates to the tx and address page (#806)
This updates the /tx and /address pages on websocket updates.
This uses stimulusjs to manage the confirmation counter on the block and address pages.
A simple pub/sub event bus is created which is used in communicating block receive
between the WebSocket update handler and the stimulus controllers. BLOCK_RECEIVED
is the main event for new blocks.
The following new address controller data-targets are added: 'numUnconfirmed',
'formattedTime', and 'txnCount'.