amount alignment fixes in chain state table (#812)
This fixes the misalignment of the amounts and the labels in the chain state table.

The decimal values in the right column were too high with respect to the label in the
left column. The align-baseline and align-bottom classes are used as needed on a
per-tr basis.

This also adds the align-baseline-rows class, specifying baseline vertical
alignment for tr elements in a table. However, the class is not used
since not all rows require the same vertical alignment.

While here, fix the disappearing space on websocket update of pool size by
moving the &nbsp; outside of the element that gets changed on websocket
update, and joining the following line.
Other fixes: The vam class was not helping on some td elements. align-bottom is
more appropriate for tr elements where multiple rows are used in the first column
(the data label). Spacing is fixed for ticket price bounds and annual stake rewards.