js: replace history state when address page goes to list-view
When the address page is loaded with no URL #anchor, the default
anchor $list-view is used in a new URL history state. This caused an
unnecessary step in the brower history and made it nearly impossible
to navigate back to say the /tx page because of the pushState.  Instead
using replaceState resolves the issue.