Modify collectionQueue to verify that new blocks are in sequence.
OnBlockConnected now relays the previous block hash from the new
block header to the collection queue for verification.
After sync, the best block height and hash are set in the collection
queue via (*collectionQueue).SetPreviousBlock.
When (*collectionQueue).ProcessBlocks verifies a block is insequence,
and calls each synchronous block connected hander, it then sets the
new block as the previous for validation of the next block.
Add (*wiredDB).GetBestBlockHeightHash to retrieve the height and hash
of the base DB's best block. Used after sync to initialize the
collection queue's previous block record.