reorg queue and shared ReorgData in txhelpers
In ntfnhandlers.go, create the basic ReorgData type without the full
chains or the common ancestor, and create a reorg signal queue
using a chan ReorgData.
Add ReorgSignaler, which processes ReorgData sent on reorgDataChan (use
queueReorg to send), allowing each package's reorg handler to finish
before proceeding to the next reorg in the queue. This should be
launched as a goroutine after starting the RPC client.
Add queueReorg to queue a new reorg via the reorgDataChan.