blockdata: update chainMonitor
Factor out data collection into (*chainMonitor).collect.
Remove the reorganizing flag since the block connected handler does not
help with the reorg anymore.
Blocks that do not connect after the last-stored block should not be
sent to BlockConnectedHandler.  It is not capable of checking that the
blocks it receives are next for dcrdata.  It only only collects data and
calles each BlockDataSaver's Store function.