sample-nginx.conf updates (#832)
This makes a number of updates to sample-nginx.conf.

Remove "http2" from port 80 server since that does not work.
Add "gzip_vary on" to allow proxies to distinguish content types.
Tweak gzip compression settings, adding more types to match recently
added assets and more.
Add a number of security related headers.  NOTE: This is a sample file,
and the sys admin is responsible for selecting appropriate values
for their server.
Add "expires modified 2d;"  to the static assets location to set the
Cache-Control and Expires headers.