explorer/views: detect unsupported address types and validate all
This recognizes additional address types, and validates all input address strings.

The following unsupported or indirectly-supported address types are now handled:

P2PK: Converted to P2PKH and a link is given to the user. This is not an
automatic redirect so that the user can see that the address has
changed.

Bitcoin: Use btcutil to attempt decoding of an address if dcrutil fails
to decode it. If it is a bitcoin address, provide a link to blockcypher.

Zero-pubkey/dummy address.

Notable code changes:

Create txhelpers.AddressValidation().
Create txhelpers types: AddressType and AddressError.
Move IsZeroHashP2PHKAddress and ValidateNetworkAddress into txhelpers.
Add Net to explorer.AddressInfo
The expStatus types now have Is* functions to easily check their value
in html templates.
Related fixes:

Full mode will not query DB if address is the zero/dummy address.
addrMetrics on the address page was a nil pointer when in lite mode.
This is fixed.