Module v4 and update depends
This starts the dcrdata module v4.

Notable, but non-breaking dependency updates:

    All of the dcrd modules that were bumped and tagged in 
(decred/dcrd@220d434) and (decred/dcrd@1a370d3)
    All of the dcrwallet modules
    github.com/asdine/storm: v2.1.1 => v2.1.2
    github.com/caarlos0/env: v3.3.0 => v3.5.0
    github.com/dgraph-io/badger: v1.5.3 => v1.5.4
    github.com/go-chi/chi: v3.3.2 => v3.3.3
    github.com/lib/pq: 20180523175426-90697d60dd84 => v1.0.0
    github.com/mattn/go-sqlite3: v1.9.0 => v1.10.0
    github.com/rs/cors: v1.5.0 => v1.6.0
    github.com/sirupsen/logrus: v1.0.5 => v1.2.0
    The golang.org/x repos: ~Sept 2018 => current

Updating badger to an untagged revision was required to eliminate 
unwelcome logging statements introduced in badger v1.5.4. A wrapper for 
badger.Logger was added to ticketpool.go so that badger logging could be 
output to the stakedb's logger. The (*Item).Value signature changed to 
no longer return the located data slice, instead performing operations 
in a closure. We define a closure that simply grabs the slice, but does 
not copy the data since the data is immediately decoded while inside the 
badger transaction.

go-socket.io and storm both pull in a number of dependencies, however 
many of them are required for tests only. Test-only packages not 
required by the main dcrdata module:

    github.com/smartystreets/assertions
    github.com/smartystreets/goconvey
    github.com/onsi/gomega
    github.com/onsi/ginko
    github.com/gopherjs/gopherjs/js
    github.com/Sereal/Sereal/Go/sereal
    github.com/DataDog/zstd

The bolt import changed from github.com/coreos/bbolt to 
go.etcd.io/bbolt, but the version is the same.