dcrsqlite: address database is locked errors
This resolves the "database is locked" issue with go-sqlite3.

"database is locked" corresponds to the `SQLITE_BUSY` error:
https://sqlite.org/rescode.html#busy

The issue is discussed at length in mattn/go-sqlite3#274. The gist is:

- SQLite should be compiled in "sequential mode" to handle multithreaded 
access. This change was made for go-sqlite v1.10.0 in 
mattn/go-sqlite3@acfa601. This dependency was updated in a previous 
dcrdata commit.

- Limiting the sql driver to just one active connection (the driver it 
maintains a connection pool) prevents concurrent access:

    db.SetMaxOpenConns(1)

- Setting a non-zero value for "busy_timeout" instructs SQLite to retry 
the DB access for the specified amount of time before throwing the 
SQLITE_BUSY error. This is a partial resolution.

    dbPath = dbPath + "?cache=shared&_busy_timeout="  + DBBusyTimeout