npm: address webpack-dev-server vulnerability
This avoids a vulnerability of in versions of webpack-dev-server prior
to 3.1.11 in which origin validation is missing. The vulnerability
allows source code to be accessed remotely. For dcrdata, which is an
open source roject, this is a minor issue or non-issue.

Ref: https://www.npmjs.com/advisories/725

This resolves the issue by updating webpack-dev-server in
devDependencies to ^3.1.14 from ^3.1.10.