allow purge of all blocks without error
DeleteBestBlock (and DeleteBlocks) returns the best block height after
attempting data removal.
Define dcrpg.NotOneRowErrMsg.
Do not return error from DeleteBlockData if deleteBlockFromChain has an
error with NotOneRowErrMsg, but do warn.
If DeleteBestBlock returns an empty string hash, that is a full rewind,
so do not return, and do add the last DeletionSummary to the result.