dcrsqlite: have StoreBlock set existing blocks to side chain first
When storing data for a main chain block, (db *DB) StoreBlock will first
set  is_mainchain=false for any other block at this height. Previously,
this was only being done in Store, which was not used during a reorg.
However, Store uses StoreBlockSummary->StoreBlock, so now this will
always happen.
Note that his is usually an unnecessary query since most of the time
StoreBlock is used outside of a reorg context.  But this solution is very
simple.  If performance of this query is an issue, it can be done
only from the ChainMonitor's reorg handler.