Load Testing Utility
This adds apiload, a simple program to implement HTTP load testing.

It uses vegeta for sending and parsing requests.

Test (a.k.a. attack) parameters are defined in a JSON-formatted definition
file, and then referenced as a command line argument or INI config file setting.

Results are output into JSON-formatted result files, as well as a summary
on the command line.

The frequency parameter, if passed will override every attacker.
Slashes at the end of config.Server and the beginning of the
endpoint are handled appropriately. Status code is 0 treated as
a dropped connection.