dcrsqlite: Only move blocks to side chain during reorg.
StoreBlock previously set all blocks at a given height as side chain
(is_mainchain=false) before adding the new block data, but this was
unnecessary unless the new block was being added as part of a reorg,
where there are actually such blocks that are being orphaned/demoted.
This change moves the setHeightToSideChain call from StoreBlock to
switchToSideChain where it belongs.
This resolves a significant performance regression with initial sync.