add merged and merged credit views
Makes a few changes to the  `AddressInfo` type. `TxCount` is an
`int` instead of a getter, and removed the `NumMergedSpent` member.
Instead, merged type rows are only counted on request, instead of
every time (see `RetrieveAddressSpentUnspent` changes). For simple
debit and credit views, this provides a substantial performance boost.

Changes the response from `explorer.AddressTable` to a JSON object, with
a `tx_count` and `html` field. The count is needed since the merged rows
are not being counted every time anymore. This is possible because if
there are credits and debits available, there are merged rows available,
an vice versa.