dcrsqlite: Add fast-sqlite-purge option.
The fast-sqlite-purge option does not follow the main chain back by N
blocks. Rather it deletes all data for blocks above a certain height
that is computed the same way as the normal purge.
Add (*WiredDB).PurgeBlocksAboveHeight and the queries:
deleteStakeInfoAboveHeightSQL and deleteBlocksAboveHeightSQL.

dcrsqlite: query is_mainchain=1 where needed

Change RetrieveLatestBlockSummary to query the best block on main chain.
Fix getHighestStakeHeight to join with the block summary table on hash
where is_mainchain=1.
Add a sanity check in PurgeBestBlock to ensure the block summary and
stake info tables are at the same height.