add live data summaries to the homepage
A bit of an experiment which adds some additional data
summary and display styles to the homepage. Introduces the
`threeSigFigs` number formatting functions in both golang template
functions and javascript `humanize` helpers, which allows short
but precise representations of values from 1e-8 up, using k,M,B
prefixes.

In the blocks section, the values of the most recent "proposed"
block are displayed prominently, along with an accounting of the
current votes for that block. In the mempool section, the count
and total DCR for "likely" transactions are displayed, along with
a small graphic that illustrates their proportions in the mempool
total. "likely" transactions do not include votes for any but the
most recently proposed block, or for duplicate votes that can pop
up when a stakeholder votes from multiple locations with a
different agenda choice. The "likely" total will count more than
20 new tickets.

All new displayed values are shown identically with or without JS
enabled in the browser. If JS is enabled, the values will update
live. Instead of simply showing the age column when JS is enabled,
the age column now replaces the timestamp column, which is redundant.
The size column is sacrificially hidden at certain breakpoints so that
the tables always display nicely. The vote type column headers are also
shortened to a single letter. A new pure-CSS tooltip is used on the
headers and elsewhere to add additional information when hovered on
desktop or tapped on mobile. Values are shown with the new
`threeSigFigs` format, which does mean that users would have to
click into the /block or /tx page to get the full 8-point precision.

Javascript `Mempool` class keeps a very shallow copy of the mempool.
No new websocket or API calls are necessary. `Mempool` has methods for
retreiving some statistical data about the mempool and tracking votes
for recent block(s).