Attempt badger db recovery when the "truncate" error is encountered
This badger's `Truncate` option to true when Open indicates it is needed.

Attempt badger db recovery when the "truncate" error is encountered.

This also removes the check for the height difference between stake db
and ticket pool db.  This will allow stake db to rewind as much as needed
to reach the ticket pool height.