integrate with other homepage changes and incorporate code suggestions.
Works in changes to master from #921, which included a full rework
of the left info column. Switches the column order based on feedback
from slack and matrix. Adds a little approved/rejected message when
the necessary votes have been received.

Switches the semantics of the vote tallying a little so that a
consensus result is indicated when available. Also changed the enum
from 0,1,2 for no, yes, missing to -1, 1, 0, which seemed
appropriate for what they represent. Added contructor for
`explorertypes.VotingInfo`.

Switched the color scheme to a simpler two-tone grey for now. Change
the `text-success` boostrap class elements to a custom `text-green`,
because the former had poor contrast on lighter backgrounds.