explorer/types: set fee and fee rate in TrimMempoolTx
This sets the following fields of the TrimmedTxInfo struct in the
TrimMempoolTx function that were not previously being set:
FormattedSize, Fee, and FeeRate. It is not clear why they were left
unset.

While in this area, also:

 - In FilterUniqueLastBlockVotes, make sure duplicate votes with the
same tx id are not added to the returned votes slice.
 - Add txhelpers.FeeRate, and use it in TxFeeRate. Add tests.
 - Fix the docs for TxFeeRate.

The changes with the TrimMempoolTx and the TrimmedTxInfo struct are
visible in a getmempooltxsResp websocket message.

Allow the now-correct fees to reach: /nexthome, explorer's websocket
"getmempooltrimmed" response, and pubsub's websocket
"getmempooltxs" response.
Also lock when reading psh.State in (*PubSubHub).receiveLoop.