explorer/types: stop the floating point math for fee summation
Assing up fees as floating point numbers was inaccurate (duh). Instead,
add as int64 and then get the decimal representation for display.

mempool: more summation in int64

ParseTxns did lots of summation in float instead of int, leading to loss
of precision.  This fixes it by summing in int64 and then converting to
coin at the end.