rebase and tweak mobile tables
Hides size and time columns on the votes table at < 576 px
screen widths. A couple other column headers are abbreviated and
given a tooltip at the same breakpoint. Better dark mode bg color
for the transaction details panel. Dropped 'Best Block' boolean
column, instead just adding indicating the best block in the
'Voting On' column.

Fixes the exchange rate info on the homepage, which was not quite
adjusted to the new layout.