dcrsqlite: add best block queries with no is_mainchain constraints
Before the sqlite table upgade is performed, it is still necessary to
query the best block height in the block and stake info tables. This
adds alternate queries that do not refernce the is_mainchain column.

Add RetrieveHighestBlockHeight and RetrieveHighestStakeHeight that
perform the query without is_mainchain. and add the wrappers
getBlockSummaryHeightAnyChain and getStakeInfoHeightAnyChain.

NewDB is modified to recognize "no such column: is_" errors, and
to fallback on the "AnyChain" functions to set the heights.

At the end of a successful table upgrade, query for the stake info best
block height using the block summary table join where is_mainchain=true.